+++ psycop/common/model_training_v2/training_method/cross_validator.py
@@ -1,0 +1,28 @@
+from psycop.common.model_training_v2.loggers.base_logger import BaselineLogger
+from psycop.common.model_training_v2.presplit_preprocessing.pipeline import (
+    PreprocessingPipeline,
+)
+from psycop.common.model_training_v2.presplit_preprocessing.polars_frame import (
+    PolarsFrame,
+)
+from psycop.common.model_training_v2.problem_type.problem_type_base import ProblemType
+from psycop.common.model_training_v2.training_method.base_training_method import (
+    TrainingMethod,
+    TrainingResult,
+)
+
+
+class CrossValidatorTrainer(TrainingMethod):
+    def __init__(
+        self,
+        data: PolarsFrame,
+        outcome_col_name: str,
+        preprocessing_pipeline: PreprocessingPipeline,
+        problem_type: ProblemType,
+        n_splits: int,
+        logger: BaselineLogger,
+    ):
+        ...
+
+    def train(self) -> TrainingResult:
+        ...
